pyproject.toml
[tool.poetry]
name = "todo-rest-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
django-rest-framework = "^0.1.0"
django-cors-headers = "^4.7.0"
strawberry-graphql-django = "^0.59.1"
strawberry-graphql = "^0.268.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
